+++ docking_md/xx.Preparation.py
@@ -2,6 +2,7 @@ import os,subprocess,wget,shutil,gzip
 from pybel import *
 
 # 0317 - modified ligand chain = protein chain
+# 0519 - modified to select manually protein pdb file
 class Protein_prep:
     def __init__(self,RosettaPath,pid,cid):
         self.RPath = RosettaPath
@@ -9,20 +10,24 @@ class Protein_prep:
         self.cid = cid
     def clean_by_rosetta(self):
         scripts_path = os.path.join(self.RPath,"tools/protein_tools/scripts/")
-
-        url = 'http://www.rcsb.org/pdb/files/%s.pdb.gz'%self.pid
-        wget.download(url)
-        with gzip.open("%s.pdb.gz"%self.pid.lower(),"rb") as F:
-            with open("%s.pdb"%self.pid,"wb") as W:
-                shutil.copyfileobj(F,W)
+        if os.path.exists("%s.pdb"%self.pid):
+                pass
+        else:
+                url = 'http://www.rcsb.org/pdb/files/%s.pdb.gz'%self.pid
+                wget.download(url)
+                with gzip.open("%s.pdb.gz"%self.pid.lower(),"rb") as F:
+                        with open("%s.pdb"%self.pid,"wb") as W:
+                                shutil.copyfileobj(F,W)
 
         subprocess.call("%s/clean_pdb.py %s %s"%(scripts_path,self.pid,self.cid),shell=True)
 
     def extract_origin_ligand(self):
         tline = []
+        print("protein: ",self.pid)
         with open("%s.pdb"%self.pid,"r") as F:
             for line in F.readlines():
                 if line.startswith("HETATM") and line[21:22] == self.cid:
+                    print(line)
                     tline.append(line.strip())
                 else:
                     pass
